Revert "AIP-72: Move Secrets Masker to task SDK (#46375)"
This reverts commit e78b5b48085e9139f70f045192f3e8169af5920d.
