+++ airflow/api_connexion/schemas/connection_schema.py
@@ -53,7 +53,7 @@ class ConnectionSchema(ConnectionCollectionItemSchema):
     def serialize_extra(obj: Connection):
         if obj.extra is None:
             return
-        from airflow.sdk.execution_time.secrets_masker import redact
+        from airflow.utils.log.secrets_masker import redact
 
         try:
             extra = json.loads(obj.extra)
