+++ airflow/api_fastapi/core_api/datamodels/assets.py
@@ -22,7 +22,7 @@ from datetime import datetime
 from pydantic import Field, field_validator
 
 from airflow.api_fastapi.core_api.base import BaseModel, StrictBaseModel
-from airflow.sdk.execution_time.secrets_masker import redact
+from airflow.utils.log.secrets_masker import redact
 
 
 class DagScheduleAssetReference(StrictBaseModel):
