+++ airflow/api_fastapi/core_api/datamodels/variables.py
@@ -23,8 +23,8 @@ from pydantic import ConfigDict, Field, model_validator
 
 from airflow.api_fastapi.core_api.base import BaseModel, StrictBaseModel
 from airflow.models.base import ID_LEN
-from airflow.sdk.execution_time.secrets_masker import redact
 from airflow.typing_compat import Self
+from airflow.utils.log.secrets_masker import redact
 
 
 class VariableResponse(BaseModel):
