+++ airflow/config_templates/config.yml
@@ -886,10 +886,10 @@ logging:
     secret_mask_adapter:
       description: |
         An import path to a function to add adaptations of each secret added with
-        ``airflow.sdk.execution_time.secrets_masker.mask_secret`` to be masked in log messages.
-        The given function is expected to require a single parameter: the secret to be adapted.
-        It may return a single adaptation of the secret or an iterable of adaptations to each be
-        masked as secrets. The original secret will be masked as well as any adaptations returned.
+        ``airflow.utils.log.secrets_masker.mask_secret`` to be masked in log messages. The given function
+        is expected to require a single parameter: the secret to be adapted. It may return a
+        single adaptation of the secret or an iterable of adaptations to each be masked as secrets.
+        The original secret will be masked as well as any adaptations returned.
       version_added: 2.6.0
       type: string
       default: ""
