+++ airflow/models/renderedtifields.py
@@ -145,7 +145,7 @@ class RenderedTaskInstanceFields(TaskInstanceDependencies):
         return prefix + ">"
 
     def _redact(self):
-        from airflow.sdk.execution_time.secrets_masker import redact
+        from airflow.utils.log.secrets_masker import redact
 
         if self.k8s_pod_yaml:
             self.k8s_pod_yaml = redact(self.k8s_pod_yaml)
