+++ airflow/utils/cli.py
@@ -35,9 +35,9 @@ import re2
 
 from airflow import settings
 from airflow.exceptions import AirflowException
-from airflow.sdk.execution_time.secrets_masker import should_hide_value_for_key
 from airflow.utils import cli_action_loggers, timezone
 from airflow.utils.log.non_caching_file_handler import NonCachingFileHandler
+from airflow.utils.log.secrets_masker import should_hide_value_for_key
 from airflow.utils.platform import getuser, is_terminal_support_colors
 
 T = TypeVar("T", bound=Callable)
