+++ docs/apache-airflow/security/secrets/mask-sensitive-values.rst
@@ -58,7 +58,7 @@ your DAG file or operator's ``execute`` function using the ``mask_secret`` funct
 
     @task
     def my_func():
-        from airflow.sdk.execution_time.secrets_masker import mask_secret
+        from airflow.utils.log.secrets_masker import mask_secret
 
         mask_secret("custom_value")
 
@@ -71,7 +71,7 @@ or
 
     class MyOperator(BaseOperator):
         def execute(self, context):
-            from airflow.sdk.execution_time.secrets_masker import mask_secret
+            from airflow.utils.log.secrets_masker import mask_secret
 
             mask_secret("custom_value")
