+++ providers/src/airflow/providers/amazon/aws/hooks/base_aws.py
@@ -58,6 +58,7 @@ from airflow.providers.amazon.aws.utils.suppress import return_on_error
 from airflow.providers_manager import ProvidersManager
 from airflow.utils.helpers import exactly_one
 from airflow.utils.log.logging_mixin import LoggingMixin
+from airflow.utils.log.secrets_masker import mask_secret
 
 BaseAwsConnection = TypeVar("BaseAwsConnection", bound=Union[boto3.client, boto3.resource])
 
@@ -67,12 +68,6 @@ if TYPE_CHECKING:
     from botocore.credentials import ReadOnlyCredentials
 
     from airflow.models.connection import Connection
-    from airflow.sdk.execution_time.secrets_masker import mask_secret
-else:
-    try:
-        from airflow.sdk.execution_time.secrets_masker import mask_secret
-    except ImportError:
-        from airflow.utils.log.secrets_masker import mask_secret
 
 _loader = botocore.loaders.Loader()
 """
