+++ providers/src/airflow/providers/amazon/aws/hooks/ecr.py
@@ -23,16 +23,10 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
+from airflow.utils.log.secrets_masker import mask_secret
 
 if TYPE_CHECKING:
     from datetime import datetime
-
-    from airflow.sdk.execution_time.secrets_masker import mask_secret
-else:
-    try:
-        from airflow.sdk.execution_time.secrets_masker import mask_secret
-    except ImportError:
-        from airflow.utils.log.secrets_masker import mask_secret
 
 logger = logging.getLogger(__name__)
